% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_missing_values.R
\name{fill.missing.values}
\alias{fill.missing.values}
\title{Fill Missing Values.}
\usage{
fill.missing.values(df, newrow.id)
}
\arguments{
\item{df}{Data Frame}

\item{newrow.id}{Row Index where the imputed row will be appended.
This will not overwrite the row that is currently at that row index.}
}
\value{
Vector with impouted values
}
\description{
Fill Missing Values.
}
\examples{
fill.missing.values(dataframe, row_index)
}
