% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_missing_dates.R
\name{find.missing.dates}
\alias{find.missing.dates}
\title{Identify Missing Dates in Time Series Data
This function returns missing dates that should be included in the df}
\usage{
find.missing.dates(d)
}
\arguments{
\item{d}{Data Frame containing time series data}
}
\value{
date object
}
\description{
Identify Missing Dates in Time Series Data
This function returns missing dates that should be included in the df
}
\examples{
find.missing.dates(data_frame)
}
