% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_row.R
\name{insertRow}
\alias{insertRow}
\title{insert a row into a data frame at a specified row index.}
\usage{
insertRow(existingDF, newrow, r)
}
\arguments{
\item{existingDF}{data frame
data frame you want to add a row to.}

\item{newrow}{integer
row index where the row should be added}

\item{r}{vector
row that you want to add to the data frame.}
}
\value{
updated data frame
}
\description{
insert a row into a data frame at a specified row index.
}
\examples{
insertRow(existingDF, newrow, r)

}
