% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/month_abv_to_num.R
\name{numMonth}
\alias{numMonth}
\title{Convert three-letter month abvs to their corresponding numbers}
\usage{
numMonth(x)
}
\arguments{
\item{x}{Three-letter month abv}
}
\value{
number between 1 and 12 (inclusive)
}
\description{
Convert three-letter month abvs to their corresponding numbers
}
\examples{
numMonth("jan")
}
